test(evaluator): split types/tests.rs into focused child modules
Issue #391 item 1: convert the oversized src/evaluator/types/tests.rs (1614
lines) into a tests/ directory module with 7 themed children plus a slim
mod.rs, mirroring the existing engine/tests and grammar/tests layouts. Pure
test code motion: the 61 evaluator::types::tests functions are unchanged
(same count before and after), including the U2 MissingPatternOperand /
RegexCompileError / is_pattern_skip coverage.

No behavior change: full suite green (1608), clippy/fmt clean both feature sets.
